requirements.txt
tmdbsimple
Pillow
